requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
Django==2.2.1
django-phonenumber-field==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.9.4
factory-boy==2.12.0
Faker==1.0.7
more-itertools==7.0.0
phonenumbers==8.10.12
pluggy==0.11.0
py==1.8.0
pytest==4.5.0
pytest-django==3.4.8
pytest-pythonpath==0.7.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
wcwidth==0.1.7
